modified based on feedback
