+++ src/server/api/controllers/boards.controller.js
@@ -21,7 +21,21 @@ const getBoardsByMemberId = async (id) => {
 
   return boards;
 };
+const getAllMembers = async (userId, boardId) => {
+  if (!Number.isInteger(Number(userId)) || !Number.isInteger(Number(boardId))) {
+    throw new InvalidIdError('Id should be an integer');
+  }
+  const members = await knex('members')
+    .join('users', 'members.userId', '=', 'users.id')
+    .select('*')
+    .where('members.boardId', boardId);
+  if (members.length === 0) {
+    throw new IncorrectEntryError(`incorrect entry with the id of ${boardId}`);
+  }
+  return members;
+};
 
 module.exports = {
   getBoardsByMemberId,
+  getAllMembers,
 };
