+++ src/server/api/routes/api-router.js
@@ -3,8 +3,6 @@ const express = require('express');
 const router = express.Router();
 
 // Router imports
-const membersRouter = require('./members.router');
-
 const usersRouter = require('./users.router');
 const swaggerJsDoc = require('swagger-jsdoc');
 const swaggerUi = require('swagger-ui-express');
@@ -31,6 +29,5 @@ router.use('/documentation', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 // Application routes
 router.use('/users', usersRouter);
-router.use('/users', membersRouter);
 
 module.exports = router;
